Split out Table and Figure
