+++ fonduer/parser/models/context.py
@@ -103,77 +103,6 @@ class TemporaryContext(object):
         raise NotImplementedError()
 
 
-class Figure(Context):
-    """A figure Context in a Document."""
-
-    __tablename__ = "figure"
-    id = Column(Integer, ForeignKey("context.id", ondelete="CASCADE"), primary_key=True)
-    document_id = Column(Integer, ForeignKey("document.id", ondelete="CASCADE"))
-    position = Column(Integer, nullable=False)
-    document = relationship(
-        "Document",
-        backref=backref("figures", order_by=position, cascade="all, delete-orphan"),
-        foreign_keys=document_id,
-    )
-    url = Column(String)
-
-    __mapper_args__ = {"polymorphic_identity": "figure"}
-
-    __table_args__ = (UniqueConstraint(document_id, position),)
-
-    def __repr__(self):
-        return "Figure(Doc: {}, Position: {}, Url: {})".format(
-            self.document.name, self.position, self.url
-        )
-
-    def __gt__(self, other):
-        # Allow sorting by comparing the string representations of each
-        return self.__repr__() > other.__repr__()
-
-
-class Cell(Context):
-    """A cell Context in a Document."""
-
-    __tablename__ = "cell"
-    id = Column(Integer, ForeignKey("context.id", ondelete="CASCADE"), primary_key=True)
-    document_id = Column(Integer, ForeignKey("document.id"))
-    table_id = Column(Integer, ForeignKey("table.id"))
-    position = Column(Integer, nullable=False)
-    document = relationship(
-        "Document",
-        backref=backref("cells", order_by=position, cascade="all, delete-orphan"),
-        foreign_keys=document_id,
-    )
-    table = relationship(
-        "Table",
-        backref=backref("cells", order_by=position, cascade="all, delete-orphan"),
-        foreign_keys=table_id,
-    )
-    row_start = Column(Integer)
-    row_end = Column(Integer)
-    col_start = Column(Integer)
-    col_end = Column(Integer)
-    html_tag = Column(Text)
-    html_attrs = Column(postgresql.ARRAY(String))
-
-    __mapper_args__ = {"polymorphic_identity": "cell"}
-
-    __table_args__ = (UniqueConstraint(document_id, table_id, position),)
-
-    def __repr__(self):
-        return "Cell(Doc: {}, Table: {}, Row: {}, Col: {}, Pos: {})".format(
-            self.document.name,
-            self.table.position,
-            tuple(set([self.row_start, self.row_end])),
-            tuple(set([self.col_start, self.col_end])),
-            self.position,
-        )
-
-    def __gt__(self, other):
-        # Allow sorting by comparing the string representations of each
-        return self.__repr__() > other.__repr__()
-
-
 class PhraseMixin(object):
     """A phrase Context in a Document."""
