+++ fonduer/parser/models/table.py
@@ -1,5 +1,6 @@
 from fonduer.parser.models.context import Context
-from sqlalchemy import Column, ForeignKey, Integer, UniqueConstraint
+from sqlalchemy import Column, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy.dialects import postgresql
 from sqlalchemy.orm import backref, relationship
 
 
@@ -22,6 +23,49 @@ class Table(Context):
 
     def __repr__(self):
         return "Table(Doc: {}, Position: {})".format(self.document.name, self.position)
+
+    def __gt__(self, other):
+        # Allow sorting by comparing the string representations of each
+        return self.__repr__() > other.__repr__()
+
+
+class Cell(Context):
+    """A cell Context in a Document."""
+
+    __tablename__ = "cell"
+    id = Column(Integer, ForeignKey("context.id", ondelete="CASCADE"), primary_key=True)
+    document_id = Column(Integer, ForeignKey("document.id"))
+    table_id = Column(Integer, ForeignKey("table.id"))
+    position = Column(Integer, nullable=False)
+    document = relationship(
+        "Document",
+        backref=backref("cells", order_by=position, cascade="all, delete-orphan"),
+        foreign_keys=document_id,
+    )
+    table = relationship(
+        "Table",
+        backref=backref("cells", order_by=position, cascade="all, delete-orphan"),
+        foreign_keys=table_id,
+    )
+    row_start = Column(Integer)
+    row_end = Column(Integer)
+    col_start = Column(Integer)
+    col_end = Column(Integer)
+    html_tag = Column(Text)
+    html_attrs = Column(postgresql.ARRAY(String))
+
+    __mapper_args__ = {"polymorphic_identity": "cell"}
+
+    __table_args__ = (UniqueConstraint(document_id, table_id, position),)
+
+    def __repr__(self):
+        return "Cell(Doc: {}, Table: {}, Row: {}, Col: {}, Pos: {})".format(
+            self.document.name,
+            self.table.position,
+            tuple(set([self.row_start, self.row_end])),
+            tuple(set([self.col_start, self.col_end])),
+            self.position,
+        )
 
     def __gt__(self, other):
         # Allow sorting by comparing the string representations of each
